requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
ruamel.yaml==0.15.37
pandas-profiling==1.4.1
jupyter==1.0.0
sphinx_rtd_theme
pandas==0.22.0
numpy==1.14.2
matplotlib==2.1.1
seaborn==0.7.1
textblob==0.15.1
textacy==0.6.0
spacy==2.0.11
dask==0.17.2